+++ services/routine/pkg/domain/aggregates/program/dailyRoutine_test.go
@@ -2,7 +2,7 @@ package program
 
 import "testing"
 
-func TestCreateDailyRoutineCausesErrWithWrongRoutineActsOrder(t *testing.T) {
+func TestCreateDailyRoutineCausingErrWithWrongRoutineActsOrder(t *testing.T) {
 	invalidRoutineActsSet := getInvalidRoutineActsSets()
 	for _, ras := range invalidRoutineActsSet {
 		_, err := CreateDailyRoutineWithoutProgramVersion(
@@ -26,7 +26,45 @@ func TestCreateDailyRoutineSettingRoutineActsRefs(t *testing.T) {
 	}
 	for _, ra := range dr.RoutineActs() {
 		if ra.DailyRoutine() != dr.Code() {
-			t.Errorf("daily routine is expected to set routine acts' reference to itself but got %v", ra.DailyRoutine())
+			t.Errorf("daily routine is expected to set routine acts' reference to itself but the reference is %v", ra.DailyRoutine())
 		}
 	}
+}
+
+func getValidDailyRoutines() DailyRoutines {
+	drs := DailyRoutines{}
+	days := []DailyRoutineDay{3, 4, 11, 19, 21}
+	for _, d := range days {
+		dr, err := CreateDailyRoutineWithoutProgramVersion(
+			d,
+			getValidRoutineActs(),
+		)
+		if err != nil {
+			panic(err)
+		}
+		drs = append(drs, dr)
+	}
+	return drs
+}
+
+func getInvalidDailyRoutinesSets() []DailyRoutines {
+	drss := []DailyRoutines{}
+	invalidDaySet := [][]DailyRoutineDay{
+		{1, 2, 2, 3, 7},
+		{2, 4, 3, 5, 9},
+		{0, 1, 2, 3, 4},
+	}
+	for _, days := range invalidDaySet {
+		drs := DailyRoutines{}
+		for _, d := range days {
+			ras := getValidRoutineActs()
+			dr, err := CreateDailyRoutineWithoutProgramVersion(d, ras)
+			if err != nil {
+				panic(err)
+			}
+			drs = append(drs, dr)
+		}
+		drss = append(drss, drs)
+	}
+	return drss
 }
